2023/10/14 添加 ddl 生成功能，调整生成相关代码结构，使用 enumeration 代替 enum 避免软件包名与关键字冲突问题

+++ src/main/kotlin/top/potmot/core/convert/TypeMapping.kt
@@ -18,7 +18,7 @@ fun jdbcTypeToJavaType(jdbcType: Int, isNotNull: Boolean = true): Class<*>? {
         Types.SMALLINT -> if (isNotNull) Short::class.java else Short::class.javaObjectType
         Types.INTEGER -> if (isNotNull) Int::class.java else Int::class.javaObjectType
         Types.BIGINT -> if (isNotNull) Long::class.java else Long::class.javaObjectType
-        Types.REAL -> if (isNotNull) Float::class.java else Long::class.javaObjectType
+        Types.REAL -> if (isNotNull) Float::class.java else Float::class.javaObjectType
         Types.FLOAT, Types.DOUBLE -> if (isNotNull) Double::class.java else Double::class.javaObjectType
         Types.DECIMAL, Types.NUMERIC -> BigDecimal::class.java
         Types.CHAR, Types.VARCHAR, Types.LONGVARCHAR, Types.NCHAR, Types.NVARCHAR, Types.LONGNVARCHAR, Types.CLOB, Types.NCLOB, Types.SQLXML, Types.DATALINK -> String::class.java
